src/main/java/dev/fadhli/playground/dailyhackerrank/algorithm/implementation/BreakingTheRecords.java
package dev.fadhli.playground.dailyhackerrank.algorithm.implementation;

import java.io.*;
import java.util.List;
import java.util.stream.Stream;

import static java.util.stream.Collectors.joining;
import static java.util.stream.Collectors.toList;

public class BreakingTheRecords {

    public static List<Integer> solution(List<Integer> scores) {
        int min = 0, max = 0;
        int lowest = scores.get(0), highest = lowest;

        for (int i = 1; i < scores.size(); i++) {
            int score = scores.get(i);
            if (score < lowest) {
                lowest = score;
                min++;
            }

            if (score > highest) {
                highest = score;
                max++;
            }
        }

        return Stream.of(max, min).collect(toList());
    }

    public static void main(String[] args) throws IOException {
        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
        BufferedWriter bufferedWriter = new BufferedWriter(new FileWriter(System.getenv("OUTPUT_PATH")));

        int n = Integer.parseInt(bufferedReader.readLine().trim());

        List<Integer> scores = Stream.of(bufferedReader.readLine().replaceAll("\\s+$", "").split(" "))
                .map(Integer::parseInt)
                .collect(toList());

        List<Integer> result = solution(scores);

        bufferedWriter.write(
                result.stream()
                        .map(Object::toString)
                        .collect(joining(" "))
                        + "\n"
        );

        bufferedReader.close();
        bufferedWriter.close();
    }

}
